requirements.txt
streamlit
groq
cohere
pypdf
langchain
langchain-community
langchain-core
langchain-groq
langchain-cohere
python-dotenv
langchain_chroma